new phase.
After last talk, we decided to compare natural MF (without histogram loss) with the results in the original paper to the consistency of the labeling.
Later we would like to check if we can observe outliers during training/post training phased.
This commit adds the data normalization and the full implementation for the raw data.